BREAKING: bump to v0.2
Redesign architecture such that more responsibilities are moved to the
Emacs side and to enable multi-process support. The code base has
been simplified on both the Emacs and R side. As a result there are some
behaviour changes and certain variables and functions have been changed
to reflect the new design.

Users using the default configuration are not affected by these changes,
but should consider adding `ess-plot-on-startup-h` to
`ess-r-post-run-hook` to enable ess-plot in all processes by default.